pyproject.toml
[project]
name = "maize-ta"
version = "0.1.0"
description = "AI Teaching Assistant Platform"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "gunicorn>=21.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "llama-index>=0.10.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "werkzeug>=3.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.45",
]
